move kit config into vite
