+++ .vscode/settings.json
@@ -32,5 +32,8 @@
 	"files.associations": {
 		"*.css": "tailwindcss"
 	},
-	"css.lint.unknownAtRules": "ignore"
+	"css.lint.unknownAtRules": "ignore",
+	"files.exclude": {
+		"**/CLAUDE.md": true
+	}
 }
