+++ lib/vite-i18n/import-map-loader/script.ts
@@ -4,7 +4,7 @@ declare const LOCALES_MAP: Record<string, string>
 declare const LOCAL_STORAGE_KEY: string
 declare const MESSAGES_MODULE_ID: string
 
-;(() => {
+{
 	const saved = localStorage.getItem(LOCAL_STORAGE_KEY)
 	const locale = saved && LOCALES.includes(saved) ? saved : BASE_LOCALE
 
@@ -16,4 +16,4 @@ declare const MESSAGES_MODULE_ID: string
 		imports: { [MESSAGES_MODULE_ID]: LOCALES_MAP[locale] },
 	})
 	document.head.appendChild(script)
-})()
+}
