+++ package.json
@@ -28,12 +28,12 @@
 		"@types/node": "^25.9.1",
 		"@types/wicg-file-system-access": "^2023.10.7",
 		"fake-indexeddb": "^6.2.5",
-		"happy-dom": "^20.9.0",
+		"happy-dom": "^20.10.1",
 		"image-size": "^2.0.2",
 		"knip": "^6.15.0",
 		"oxfmt": "^0.53.0",
 		"svelte": "5.56.1",
-		"svelte-check": "4.5.0",
+		"svelte-check": "4.6.0",
 		"tailwindcss": "^4.3.0",
 		"typescript": "6.0.3",
 		"unplugin-auto-import": "^21.0.0",
