+++ tsconfig.node.json
@@ -3,6 +3,6 @@
 	"compilerOptions": {
 		"types": ["@types/node"]
 	},
-	"files": ["vite.config.ts", "svelte.config.ts", "src/lib/theme.ts", ".generated/i18n/runtime.ts"],
+	"files": ["vite.config.ts", "src/lib/theme.ts", ".generated/i18n/runtime.ts"],
 	"include": ["lib/**/*", "scripts/**/*"]
 }
